Introduce Sleef as an experimental uKernel library (#32)
Adds machinery to build + install sleef from inside the triton plugin. This is complicated by a few things - for one, we don't have top-level CMake project control. For another, the typical Triton install process is to use pip, not cmake commands directly. Each pip command re-runs configure, which seems to pass a temporary ninja path to cmake. If the cmake configure is re-used, the old ninja path may be cached which will cause a failure. For top-level includes we can force the new ninja path, but for their includes we have to get creative. Sleef relies on the submodule project tlfloat, so I had to get creative and force sleef to re-checkout tlfloat for each pip/cmake re-run.