[NFC]: Refactor load/store LLVM code into Utilities (#40)
NFC from the perspective of the confined space of tutorial 1, but I discovered a bug in the previous version of the code that was "over-vectorizing", so this initially lost performance. Enabling proper "getVectorSize' for the NPU version of LoadStoreOpToLLVM returned us to previous performance with the more correct masking code.